+++ Sources/MarkersExtractor/Export/ExportProfile/ExportProfileFormat.swift
@@ -12,8 +12,10 @@ public enum ExportProfileFormat: String {
     case compressor
     case csv
     case json
+    case markdown
     case midi
     case notion
+    case srt
     case tsv
     case xlsx
     case youtube
@@ -40,6 +42,8 @@ extension ExportProfileFormat {
         case .compressor: "Compressor Chapters"
         case .csv: "CSV"
         case .json: "JSON"
+        case .markdown: "Markdown"
+        case .srt: "SubRip (SRT)"
         case .midi: "MIDI File"
         case .notion: "Notion"
         case .tsv: "TSV"
@@ -54,6 +58,8 @@ extension ExportProfileFormat {
         case .compressor: CompressorProfile.self
         case .csv: CSVProfile.self
         case .json: JSONProfile.self
+        case .markdown: MarkdownProfile.self
+        case .srt: SubRipProfile.self
         case .midi: MIDIFileExportProfile.self
         case .notion: NotionExportProfile.self
         case .tsv: TSVProfile.self