+++ Sources/MarkersExtractor/Export/Result/ExportResult.swift
@@ -28,6 +28,12 @@ public struct ExportResult {
     /// Plain Text manifest file path, if applicable to the profile. `nil` if not applicable.
     public var txtManifestPath: URL?
     
+    /// Markdown manifest file path, if applicable to the profile. `nil` if not applicable.
+    public var mdManifestPath: URL?
+    
+    /// SubRip manifest file path, if applicable to the profile. `nil` if not applicable.
+    public var srtManifestPath: URL?
+    
     /// JSON manifest file path, if applicable to the profile. `nil` if not applicable.
     public var jsonManifestPath: URL?
     
@@ -47,6 +53,8 @@ public struct ExportResult {
         csvManifestPath: URL? = nil,
         tsvManifestPath: URL? = nil,
         txtManifestPath: URL? = nil,
+        mdManifestPath: URL? = nil,
+        srtManifestPath: URL? = nil,
         jsonManifestPath: URL? = nil,
         midiFilePath: URL? = nil,
         xlsxManifestPath: URL? = nil
@@ -58,6 +66,8 @@ public struct ExportResult {
         self.csvManifestPath = csvManifestPath
         self.tsvManifestPath = tsvManifestPath
         self.txtManifestPath = txtManifestPath
+        self.mdManifestPath = txtManifestPath
+        self.srtManifestPath = srtManifestPath
         self.jsonManifestPath = jsonManifestPath
         self.midiFilePath = midiFilePath
         self.xlsxManifestPath = xlsxManifestPath
@@ -94,6 +104,12 @@ extension ExportResult {
         if let value = dict[.txtManifestPath], case let .url(url) = value {
             txtManifestPath = url
         }
+        if let value = dict[.mdManifestPath], case let .url(url) = value {
+            mdManifestPath = url
+        }
+        if let value = dict[.srtManifestPath], case let .url(url) = value {
+            srtManifestPath = url
+        }
         if let value = dict[.jsonManifestPath], case let .url(url) = value {
             jsonManifestPath = url
         }
@@ -122,6 +138,8 @@ extension ExportResult {
         dict[.csvManifestPath] = csvManifestPath?.path
         dict[.tsvManifestPath] = tsvManifestPath?.path
         dict[.txtManifestPath] = txtManifestPath?.path
+        dict[.mdManifestPath] = mdManifestPath?.path
+        dict[.srtManifestPath] = srtManifestPath?.path
         dict[.jsonManifestPath] = jsonManifestPath?.path
         dict[.midiFilePath] = midiFilePath?.path
         dict[.xlsxManifestPath] = xlsxManifestPath?.path