+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # MarkersExtractor Change Log
 
-## [0.3.16](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.16) (2025-07-??)
+## [0.3.17](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.17) (2025-07-??)
+
+### New Features
+
+- Added SubRip export profile to export SRT subtitles (#128)
+
+## [0.3.16](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.16) (2025-07-07)
 
 ### New Features
 
 - Added Apple Compressor export profile to export chapter marker lists (#123)
+- Added Markdown export profile (#127)
 
 ### Maintenance